Move some stuff around, add return statements
