+++ src/com/statements/If.java
@@ -1,8 +1,6 @@
 package com.statements;
 
-import com.Block;
 import com.expressions.Expression;
-import com.statements.Statement;
 
 public class If extends Statement {
     public Expression condition;
